requirements.txt
python-telegram-bot[ext]>=20.7
motor>=3.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
dnspython>=2.4.2
